requirements.txt
dash==2.6.0
pandas
numpy
gunicorn
dash_bootstrap_components
dash_html_components
dash_core_components
dash-auth
pathlib
plotly.graph_objs
